feat: rename MainNavigator to NavBar
