+++ sites/tutors-course-reader/src/components/navigators/NavBar.svelte
